codecheck

+++ rosen/modules/render_service/core/feature/hdr/hetero_hdr/rs_hetero_hdr_util.cpp
@@ -46,10 +46,10 @@ void RSHeteroHDRUtil::GenDrawHDRBufferParams(const DrawableV2::RSSurfaceRenderNo
     if (!isFixedDstBuffer) {
         auto layerSrc = surfaceParams->GetLayerInfo().srcRect;
         Drawing::Matrix matrixLoc;
-        auto sx = float(layerSrc.w) / hpaeDstRect.right;
-        auto sy = float(layerSrc.h) / hpaeDstRect.bottom;
-        auto tx = float(layerSrc.x);
-        auto ty = float(layerSrc.y);
+        auto sx = static_cast<float>(layerSrc.w) / static_cast<float>(hpaeDstRect.right);
+        auto sy = static_cast<float>(layerSrc.h) / static_cast<float>(hpaeDstRect.bottom);
+        auto tx = static_cast<float>(layerSrc.x);
+        auto ty = static_cast<float>(layerSrc.y);
         matrixLoc.SetScaleTranslate(sx, sy, tx, ty);
         localParam.matrix = drawableParams.matrix;
         localParam.matrix.PreConcat(matrixLoc);
